@@ -67,18 +57,27 @@ ext {
 }
 
 java {
-    disableAutoTargetJvm()
+    sourceCompatibility = 17
 }
 
-apply plugin: 'kotlin'
-apply plugin: 'eclipse'
-apply plugin: 'org.springframework.boot'
+compileKotlin {
+    kotlinOptions {
+        freeCompilerArgs = ["-Xjsr305=strict"]
+        jvmTarget = "17"
+    }
+}
 
 jar {
     archiveBaseName.set('alfio-pi')
-    version = "${project.version}"
+    archiveVersion.set("${project.version}")
 }
 jar.dependsOn(":frontend:jar")
+
+bootJar {
+    archiveBaseName.set('alfio-pi')
+    archiveVersion.set("${project.version}")
+    archiveClassifier.set("boot")
+}
 
 processResources.doLast {
 
