@@ -20,11 +20,9 @@ package alfio.pi.manager
 import alfio.pi.model.LabelLayout
 import alfio.pi.model.Ticket
 import com.google.gson.Gson
-import org.junit.Assert.*
-import org.junit.Test
+import org.junit.jupiter.api.Assertions.*
+import org.junit.jupiter.api.Test
 import org.mockito.Mockito
-import java.nio.file.Files
-import java.nio.file.Paths
 import java.util.*
 
 class LabelManagerTest {
@@ -54,6 +52,7 @@ class LabelManagerTest {
     fun testGenerateLabelDymoWithThreeAdditionalRows() {
         var bytes = generatePDFLabel("George", "William", listOf("First", "Second", "Third"), "12345678", UUID.randomUUID().toString(), "12345678", null,false).invoke(DymoLW450Turbo41x89())
         assertTrue(bytes.isNotEmpty())
+        // Files.write(Files.createTempFile("test", ".pdf"), bytes)
         bytes = generatePDFLabel("George", "William", listOf("First", "Second", "Third"), "12345678", UUID.randomUUID().toString(), "12345678", null, true).invoke(DymoLW450Turbo41x89())
         assertTrue(bytes.isNotEmpty())
     }
